requirements.txt
openai>=1.0.0
anthropic>=0.18.0
pydub>=0.25.1
watchdog>=3.0.0
rq>=1.15.0
redis>=4.5.0
notion-client>=2.0.0
python-dotenv>=1.0.0
